Improved "wait" implementation
Tracks updates in the "changes" table for, keeping the last 1000 changes.  When
a node is modified or deleted the old row has its "deleted" column set to
the current index, and a row with the new value is inserted with "deleted" set
to 0.

Starts a go routine to poll the changes table, and keep a local buffer of the
change information. When a new "wait" request is received, it waits for a
matching update from the changes table.

This follows the real etcd wait implementation more closely for observing
recent history without missing changes.