Add flags for advertizing client URLs
This adds flags -listen-client-urls and -advertise-client-urls with the same
purpose as the regular etcd server.

-listen-client-urls contains one or more addresses for the server to listen on
-advertise-client-urls contains the corresponding public URLs for clients to
connect to

To be compatible with clients like etcdctl, the etcdb server needs to advertize
its own publicly accessible address when asked. For example, to run in Docker
"-advertise-client-urls" needs to contain the public IP and port that outside
clients will connect to.